change layout to be more in center of the main app
